+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Changes to the plugin.
 - References in a multireplace test are now handled correctly. (reported by nocturno)
 - Multireplaces with references now have proper syntax highlighting.
 - A `*goto_scene` statement with a variable reference for the scene is no longer flagged as an error. (reported by nocturno)
+- A `*goto_scene` statement with a variable reference for the label now correctly indexes the reference's location. (reported by nocturno)
 - Minor fix to word counting in options with a multireplace.
 
 ## [2.3.0] - 2021-08-14